[Add]星評価

+++ db/migrate/20231005041738_create_souvenirs.rb
@@ -5,7 +5,7 @@ class CreateSouvenirs < ActiveRecord::Migration[6.1]
       t.string  :souvenir_name
       t.text    :introduction
       t.integer :category
-      t.integer :average_star
+      t.float :average_star
       t.timestamps
     end
   end
